% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estTb.R
\name{estTb}
\alias{estTb}
\title{Predict brightness temperatures}
\usage{
estTb(tbH, tbV, fH, fV, gamma, Tair, Tsoil, omega, rhfac)
}
\arguments{
\item{fH}{Estimated horizontal fresnel reflectivity}

\item{fV}{Estimated vertical fresnel reflectivity}

\item{rhfac}{}

\item{TbH}{Observed horizontally polarized brightness temperature}

\item{TbV}{Observed vertically polarized brightness temperature}
}
\value{
Vertical and horizontally polarized brightness temperatures
}
\description{
Predict brightness temperatures
}
