% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fresnelr.R
\name{fresnelr}
\alias{fresnelr}
\title{Fresnel's reflectivity calculation}
\usage{
fresnelr(eps, theta, h)
}
\arguments{
\item{eps}{Dielectric}

\item{theta}{Incidence angle in degrees}

\item{h}{Soil surface roughness}
}
\value{
List of horizontal and vertical polarization reflectivity
}
\description{
Fresnel's reflectivity calculation
}
