% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sin.R
\name{gen_sin}
\alias{gen_sin}
\title{Generate sinusoid}
\usage{
gen_sin(len, rangeL, rangeH)
}
\arguments{
\item{len}{length of the output vector}

\item{rangeL}{lower bound to values}

\item{rangeH}{upper bound to values}
}
\description{
Generate sinusoid
}
