% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSMVOD.R
\name{get_sm_vod}
\alias{get_sm_vod}
\title{Title}
\usage{
get_sm_vod(
  smc,
  TbH,
  TbV,
  Tair,
  Tsoil = NULL,
  vod = seq(0, 3, length.out = 10),
  omega = 0.05,
  inc_angle = 40,
  cf = 0.232,
  roughness = 0.16,
  twotemps = F,
  split = F,
  rtms = FALSE,
  mpdi = FALSE
)
}
\arguments{
\item{mpdi}{}
}
\description{
Title
}
