% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mironov.R
\name{mironov}
\alias{mironov}
\title{mironov}
\usage{
mironov(f, smv, cf)
}
\arguments{
\item{f}{Frequency}

\item{smv}{Soil moisture volume}

\item{cf}{Clay fraction of soil}
}
\value{
A list of the real and imaginary parts of the dielectric constant
}
\description{
Calculate the dielectric constant of soil
}
