% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optomegarough.R
\name{optOR}
\alias{optOR}
\title{Optimize omega and roughness}
\usage{
optOR(df, range_omega, range_roughness, precision)
}
\arguments{
\item{df}{A data.frame}

\item{range_omega}{Range of test omega values}

\item{range_roughness}{Range of test roughness values}

\item{precision}{Numeric indicating increment step of the grid search}
}
\value{
List of all residuals and the best residuals
}
\description{
Optimize omega and roughness
}
