% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retVOD.R
\name{print.retVOD}
\alias{print.retVOD}
\title{Print retVOD object}
\usage{
\method{print}{retVOD}(x, ...)
}
\arguments{
\item{x}{Object of class `retVOD`}

\item{...}{Not used.}
}
\value{
Printed retrieval results
}
\description{
Print retVOD object
}
