% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retVOD.R
\name{ret_vod_dual}
\alias{ret_vod_dual}
\title{Title}
\usage{
ret_vod_dual(
  TbH,
  TbV,
  sm,
  vod = seq(0, 3, by = 0.01),
  Tair,
  Tsoil = NULL,
  omega = 0.05,
  roughness = 0.15,
  inc_angle = 40,
  ret_fx = get_sm_vod,
  twotemps = TRUE,
  mpdi = F,
  split = F,
  rtms = F
)
}
\arguments{
\item{rtms}{}
}
\description{
Title
}
