% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solveSmVod.R
\name{solveSmVod}
\alias{solveSmVod}
\title{Solve for soil moisture and VOD}
\usage{
solveSmVod(reflecs, gamma, tbH, tbV, Tair, Tsoil, omega, mat = F)
}
\arguments{
\item{reflecs}{H and V pol reflectivities}

\item{gamma}{Gamma estimates for each test VOD value}

\item{tbH}{Horizontal brightness temperature}

\item{tbV}{Vertical brightness temperature}

\item{Tair}{Air temperature}

\item{Tsoil}{Soil temperature}

\item{omega}{Scattering albedo}

\item{mat}{TRUE returns the cost function matrix}
}
\value{
List of predicted brightness temperatures, soil moisture and VOD:
}
\description{
This function uses the a range of input soil moisture and vod values to solve for the best VOD value given the air, soil, and observed brightness temperatures.
}
\details{
The function returns the predicted brightness temperatures (i.e., using tau-omega), residuals for each polarization, and the predicted soil moisture and VOD. Mironov is used to determine the dielectric constant for a given soil moisture and has a set frequency of 1.4e9Hz (for L-band). Also the clay fraction at MOFLUX is 23.2%.
}
