% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tauomega.R
\name{tau_omega}
\alias{tau_omega}
\title{Tau-omega model}
\usage{
tau_omega(r, rhfac, gamma, Tair = NULL, Tsoil = NULL, omega, emiss = F)
}
\arguments{
\item{r}{Fresnel reflectivity}

\item{rhfac}{Roughness factor to apply to `r`}

\item{gamma}{Vegetation transmissivity transfer function. VOD divided by the cosine of incidence angle}

\item{Tair}{Air temperature in deg Kelvin}

\item{Tsoil}{Soil temperature in deg Kelvin}

\item{omega}{Scattering albedo}

\item{emiss}{Do you want to only calculate the emissitivity?}
}
\value{
Vector of brightness temperatures
}
\description{
Tau-omega model
}
