% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_residuals.R
\name{tbResiduals}
\alias{tbResiduals}
\title{Calculate residuals for predicted brightness temperatures}
\usage{
tbResiduals(tbH, tbV, tbHpred, tbVpred)
}
\arguments{
\item{tbH}{Obs.H polarization brightness temperature}

\item{tbV}{Obs.V polarization brightness temperature}

\item{tbHpred}{Pred. H polarization brightness temperature}

\item{tbVpred}{Pred. V polarization brightness temperature}
}
\value{
Residuals for each predicted brightness temperature
}
\description{
Calculate residuals for predicted brightness temperatures
}
