% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estTb.R
\name{estTb}
\alias{estTb}
\title{Predict brightness temperatures}
\usage{
estTb(tbH, tbV, fH, fV, gamma, Tair, Tsoil, omega, rhfac)
}
\arguments{
\item{tbH}{Observed horizontally polarized brightness temperature}

\item{tbV}{Observed vertically polarized brightness temperature}

\item{fH}{Estimated horizontal fresnel reflectivity}

\item{fV}{Estimated vertical fresnel reflectivity}

\item{gamma}{Transfer function}

\item{Tair}{Air temperature}

\item{Tsoil}{Soil temperature}

\item{omega}{Canopy opacity}

\item{rhfac}{Roughness factor}
}
\value{
Vertical and horizontally polarized brightness temperatures
}
\description{
Predict brightness temperatures
}
